+++ lib/plugins/plugins.php
@@ -6,14 +6,15 @@
  * It is expected that theme authors would copy and paste this code into their
  * functions.php file, and amend to suit.
  *
+ * @see http://tgmpluginactivation.com/configuration/ for detailed documentation.
+ *
  * @package    TGM-Plugin-Activation
  * @subpackage Example
- * @version    2.4.0
- * @author     Thomas Griffin <thomasgriffinmedia.com>
- * @author     Gary Jones <gamajo.com>
- * @copyright  Copyright (c) 2014, Thomas Griffin
+ * @version    2.5.2
+ * @author     Thomas Griffin, Gary Jones, Juliette Reinders Folmer
+ * @copyright  Copyright (c) 2011, Thomas Griffin
  * @license    http://opensource.org/licenses/gpl-2.0.php GPL v2 or later
- * @link       https://github.com/thomasgriffin/TGM-Plugin-Activation
+ * @link       https://github.com/TGMPA/TGM-Plugin-Activation
  */
 
 /**
@@ -25,8 +26,10 @@ add_action( 'tgmpa_register', 'wsm_register_required_plugins' );
 /**
  * Register the required plugins for this theme.
  *
- * In this example, we register two plugins - one included with the TGMPA library
- * and one from the .org repo.
+ * In this example, we register five plugins:
+ * - one included with the TGMPA library
+ * - two from an external source, one from an arbitrary source, one from a GitHub repository
+ * - two from the .org repo, where one demonstrates the use of the `is_callable` argument
  *
  * The variable passed to tgmpa_register_plugins() should be an array of plugin
  * arrays.
@@ -35,69 +38,127 @@ add_action( 'tgmpa_register', 'wsm_register_required_plugins' );
  * TGM_Plugin_Activation class constructor.
  */
 function wsm_register_required_plugins() {
-
-	/**
+	/*
 	 * Array of plugin arrays. Required keys are name and slug.
 	 * If the source is NOT from the .org repo, then source is also required.
 	 */
 	$plugins = array(
 
-		// This is an example of how to include a plugin pre-packaged with a theme
+		// This is an example of how to include a plugin bundled with a theme.
 		array(
-			'name'     				=> 'Soliloquy', // The plugin name
-			'slug'     				=> 'soliloquy', // The plugin slug (typically the folder name)
-			'source'   				=> 'https://s3.amazonaws.com/wsm_files/plugins/soliloquy.zip', // The plugin source
-			'required' 				=> false, // If false, the plugin is only 'recommended' instead of required
-			'force_activation' 		=> false, // If true, plugin is activated upon theme activation and cannot be deactivated until theme switch
-			'force_deactivation' 	=> false, // If true, plugin is deactivated upon theme switch, useful for theme-specific plugins
-            'external_url'       => 'http://www.web-savvy-marketing.com/go/soliloquy/', // If set, overrides default API URL and points to an external URL.
+			'name'               => 'Soliloquy',
+			'slug'               => 'soliloquy',
+			'source'             => 'https://s3.amazonaws.com/wsm_files/plugins/soliloquy.zip',
+			'required'           => false,
+			'version'            => '',
+			'force_activation'   => false,
+			'force_deactivation' => false,
+			'external_url'       => 'http://www.web-savvy-marketing.com/go/soliloquy/',
+			'is_callable'        => '',
 		),
 
 	);
 
-	// Change this to your theme text domain, used for internationalising strings
 	$theme_text_domain = 'deborah';
 
-	/**
+	/*
 	 * Array of configuration settings. Amend each line as needed.
-	 * If you want the default strings to be available under your own theme domain,
-	 * leave the strings uncommented.
-	 * Some of the strings are added into a sprintf, so see the comments at the
-	 * end of each line for what each argument will be.
+	 *
+	 * TGMPA will start providing localized text strings soon. If you already have translations of our standard
+	 * strings available, please help us make TGMPA even better by giving us access to these translations or by
+	 * sending in a pull-request with .po file(s) with the translations.
+	 *
+	 * Only uncomment the strings in the config array if you want to customize the strings.
 	 */
 	$config = array(
-		'domain'       		=> $theme_text_domain,         	// Text domain - likely want to be the same as your theme.
-		'default_path' 		=> '',                         	// Default absolute path to pre-packaged plugins
-		'parent_menu_slug' 	=> 'themes.php', 				// Default parent menu slug
-		'parent_url_slug' 	=> 'themes.php', 				// Default parent URL slug
-		'menu'         		=> 'install-required-plugins', 	// Menu slug
-		'has_notices'      	=> true,                       	// Show admin notices or not
-        'dismissable'  		=> true,                    	// If false, a user cannot dismiss the nag message.
-        'dismiss_msg'  		=> '',                     		// If 'dismissable' is false, this message will be output at top of nag.
-		'is_automatic'    	=> false,					   	// Automatically activate plugins after installation or not
-		'message' 			=> '',							// Message to output right before the plugins table
-		'strings'      		=> array(
-			'page_title'                       			=> __( 'Install Required Plugins', $theme_text_domain ),
-			'menu_title'                       			=> __( 'Install Plugins', $theme_text_domain ),
-			'installing'                       			=> __( 'Installing Plugin: %s', $theme_text_domain ), // %1$s = plugin name
-			'oops'                             			=> __( 'Something went wrong with the plugin API.', $theme_text_domain ),
-			'notice_can_install_required'     			=> _n_noop( 'This theme requires the following plugin: %1$s.', 'This theme requires the following plugins: %1$s.' ), // %1$s = plugin name(s)
-			'notice_can_install_recommended'			=> _n_noop( 'This theme recommends the following plugin: %1$s.', 'This theme recommends the following plugins: %1$s.' ), // %1$s = plugin name(s)
-			'notice_cannot_install'  					=> _n_noop( 'Sorry, but you do not have the correct permissions to install the %s plugin. Contact the administrator of this site for help on getting the plugin installed.', 'Sorry, but you do not have the correct permissions to install the %s plugins. Contact the administrator of this site for help on getting the plugins installed.' ), // %1$s = plugin name(s)
-			'notice_can_activate_required'    			=> _n_noop( 'The following required plugin is currently inactive: %1$s.', 'The following required plugins are currently inactive: %1$s.' ), // %1$s = plugin name(s)
-			'notice_can_activate_recommended'			=> _n_noop( 'The following recommended plugin is currently inactive: %1$s.', 'The following recommended plugins are currently inactive: %1$s.' ), // %1$s = plugin name(s)
-			'notice_cannot_activate' 					=> _n_noop( 'Sorry, but you do not have the correct permissions to activate the %s plugin. Contact the administrator of this site for help on getting the plugin activated.', 'Sorry, but you do not have the correct permissions to activate the %s plugins. Contact the administrator of this site for help on getting the plugins activated.' ), // %1$s = plugin name(s)
-			'notice_ask_to_update' 						=> _n_noop( 'The following plugin needs to be updated to its latest version to ensure maximum compatibility with this theme: %1$s.', 'The following plugins need to be updated to their latest version to ensure maximum compatibility with this theme: %1$s.' ), // %1$s = plugin name(s)
-			'notice_cannot_update' 						=> _n_noop( 'Sorry, but you do not have the correct permissions to update the %s plugin. Contact the administrator of this site for help on getting the plugin updated.', 'Sorry, but you do not have the correct permissions to update the %s plugins. Contact the administrator of this site for help on getting the plugins updated.' ), // %1$s = plugin name(s)
-			'install_link' 					  			=> _n_noop( 'Begin installing plugin', 'Begin installing plugins' ),
-			'activate_link' 				  			=> _n_noop( 'Activate installed plugin', 'Activate installed plugins' ),
-			'return'                           			=> __( 'Return to Required Plugins Installer', $theme_text_domain ),
-			'plugin_activated'                 			=> __( 'Plugin activated successfully.', $theme_text_domain ),
-			'complete' 									=> __( 'All plugins installed and activated successfully. %s', $theme_text_domain ), // %1$s = dashboard link
-			'nag_type'									=> 'updated', // Determines admin notice type - can only be 'updated' or 'error'
-		)
+		'id'           => 'wsm',                   // Unique ID for hashing notices for multiple instances of TGMPA.
+		'default_path' => '',                      // Default absolute path to bundled plugins.
+		'menu'         => 'wsm-install-plugins',   // Menu slug.
+		'parent_slug'  => 'themes.php',            // Parent menu slug.
+		'capability'   => 'edit_theme_options',    // Capability needed to view plugin install page, should be a capability associated with the parent menu used.
+		'has_notices'  => true,                    // Show admin notices or not.
+		'dismissable'  => true,                    // If false, a user cannot dismiss the nag message.
+		'dismiss_msg'  => '',                      // If 'dismissable' is false, this message will be output at top of nag.
+		'is_automatic' => false,                   // Automatically activate plugins after installation or not.
+		'message'      => '',                      // Message to output right before the plugins table.
+
+		'strings'      => array(
+			'page_title'                      => __( 'Install Required Plugins', $theme_text_domain ),
+			'menu_title'                      => __( 'Install Plugins', $theme_text_domain ),
+			'installing'                      => __( 'Installing Plugin: %s', $theme_text_domain ), // %s = plugin name.
+			'oops'                            => __( 'Something went wrong with the plugin API.', $theme_text_domain ),
+			'notice_can_install_required'     => _n_noop(
+				'This theme requires the following plugin: %1$s.',
+				'This theme requires the following plugins: %1$s.',
+				$theme_text_domain
+			), // %1$s = plugin name(s).
+			'notice_can_install_recommended'  => _n_noop(
+				'This theme recommends the following plugin: %1$s.',
+				'This theme recommends the following plugins: %1$s.',
+				$theme_text_domain
+			), // %1$s = plugin name(s).
+			'notice_cannot_install'           => _n_noop(
+				'Sorry, but you do not have the correct permissions to install the %1$s plugin.',
+				'Sorry, but you do not have the correct permissions to install the %1$s plugins.',
+				$theme_text_domain
+			), // %1$s = plugin name(s).
+			'notice_ask_to_update'            => _n_noop(
+				'The following plugin needs to be updated to its latest version to ensure maximum compatibility with this theme: %1$s.',
+				'The following plugins need to be updated to their latest version to ensure maximum compatibility with this theme: %1$s.',
+				$theme_text_domain
+			), // %1$s = plugin name(s).
+			'notice_ask_to_update_maybe'      => _n_noop(
+				'There is an update available for: %1$s.',
+				'There are updates available for the following plugins: %1$s.',
+				$theme_text_domain
+			), // %1$s = plugin name(s).
+			'notice_cannot_update'            => _n_noop(
+				'Sorry, but you do not have the correct permissions to update the %1$s plugin.',
+				'Sorry, but you do not have the correct permissions to update the %1$s plugins.',
+				$theme_text_domain
+			), // %1$s = plugin name(s).
+			'notice_can_activate_required'    => _n_noop(
+				'The following required plugin is currently inactive: %1$s.',
+				'The following required plugins are currently inactive: %1$s.',
+				$theme_text_domain
+			), // %1$s = plugin name(s).
+			'notice_can_activate_recommended' => _n_noop(
+				'The following recommended plugin is currently inactive: %1$s.',
+				'The following recommended plugins are currently inactive: %1$s.',
+				$theme_text_domain
+			), // %1$s = plugin name(s).
+			'notice_cannot_activate'          => _n_noop(
+				'Sorry, but you do not have the correct permissions to activate the %1$s plugin.',
+				'Sorry, but you do not have the correct permissions to activate the %1$s plugins.',
+				$theme_text_domain
+			), // %1$s = plugin name(s).
+			'install_link'                    => _n_noop(
+				'Begin installing plugin',
+				'Begin installing plugins',
+				$theme_text_domain
+			),
+			'update_link' 					  => _n_noop(
+				'Begin updating plugin',
+				'Begin updating plugins',
+				$theme_text_domain
+			),
+			'activate_link'                   => _n_noop(
+				'Begin activating plugin',
+				'Begin activating plugins',
+				$theme_text_domain
+			),
+			'return'                          => __( 'Return to Required Plugins Installer', $theme_text_domain ),
+			'plugin_activated'                => __( 'Plugin activated successfully.', $theme_text_domain ),
+			'activated_successfully'          => __( 'The following plugin was activated successfully:', $theme_text_domain ),
+			'plugin_already_active'           => __( 'No action taken. Plugin %1$s was already active.', $theme_text_domain ),  // %1$s = plugin name(s).
+			'plugin_needs_higher_version'     => __( 'Plugin not activated. A higher version of %s is needed for this theme. Please update the plugin.', $theme_text_domain ),  // %1$s = plugin name(s).
+			'complete'                        => __( 'All plugins installed and activated successfully. %1$s', $theme_text_domain ), // %s = dashboard link.
+			'contact_admin'                   => __( 'Please contact the administrator of this site for help.', 'tgmpa' ),
+
+			'nag_type'                        => 'updated', // Determines admin notice type - can only be 'updated', 'update-nag' or 'error'.
+		),
+
 	);
 
 	tgmpa( $plugins, $config );
-
 }